rename consumer_group module to consumer.rs
