--- foreign/cpp/src/consumer_group.rs
@@ -1,70 +1,0 @@
-// Licensed to the Apache Software Foundation (ASF) under one
-// or more contributor license agreements.  See the NOTICE file
-// distributed with this work for additional information
-// regarding copyright ownership.  The ASF licenses this file
-// to you under the Apache License, Version 2.0 (the
-// "License"); you may not use this file except in compliance
-// with the License.  You may obtain a copy of the License at
-//
-//   http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing,
-// software distributed under the License is distributed on an
-// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-// KIND, either express or implied.  See the License for the
-// specific language governing permissions and limitations
-// under the License.
-
-use crate::ffi;
-use iggy::prelude::ConsumerGroupDetails as RustConsumerGroupDetails;
-use iggy_common::{
-    ConsumerGroup as RustConsumerGroup, ConsumerGroupInfo as RustConsumerGroupInfo,
-    ConsumerGroupMember as RustConsumerGroupMember,
-};
-
-impl From<RustConsumerGroupInfo> for ffi::ConsumerGroupInfo {
-    fn from(group: RustConsumerGroupInfo) -> Self {
-        ffi::ConsumerGroupInfo {
-            stream_id: group.stream_id,
-            topic_id: group.topic_id,
-            group_id: group.group_id,
-        }
-    }
-}
-
-impl From<RustConsumerGroupMember> for ffi::ConsumerGroupMember {
-    fn from(member: RustConsumerGroupMember) -> Self {
-        ffi::ConsumerGroupMember {
-            id: member.id,
-            partitions_count: member.partitions_count,
-            partitions: member.partitions,
-        }
-    }
-}
-
-impl From<RustConsumerGroup> for ffi::ConsumerGroup {
-    fn from(group: RustConsumerGroup) -> Self {
-        ffi::ConsumerGroup {
-            id: group.id,
-            name: group.name,
-            partitions_count: group.partitions_count,
-            members_count: group.members_count,
-        }
-    }
-}
-
-impl From<RustConsumerGroupDetails> for ffi::ConsumerGroupDetails {
-    fn from(group: RustConsumerGroupDetails) -> Self {
-        ffi::ConsumerGroupDetails {
-            id: group.id,
-            name: group.name,
-            partitions_count: group.partitions_count,
-            members_count: group.members_count,
-            members: group
-                .members
-                .into_iter()
-                .map(ffi::ConsumerGroupMember::from)
-                .collect(),
-        }
-    }
-}
